add coming soon page for blog
