+++ app/blogs/page.tsx
@@ -1,212 +1,12 @@
-"use client";
+import BlogsComingSoon from '@/features/BlogsComingSoon'
+import React from 'react'
 
-import { useState } from "react";
-import { motion } from "framer-motion";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
-import {
-  Sheet,
-  SheetContent,
-  SheetHeader,
-  SheetTitle,
-  SheetTrigger,
-} from "@/components/ui/sheet";
-import { Code2, Search, ChevronRight, Menu } from "lucide-react";
-import Link from "next/link";
-import { blogPosts } from "@/constants/blog-posts";
-import { useRouter } from "next/navigation";
-
-const fadeIn = {
-  initial: { opacity: 0, y: 20 },
-  animate: { opacity: 1, y: 0 },
-  transition: { duration: 0.6 },
-};
-
-const stagger = {
-  animate: {
-    transition: {
-      staggerChildren: 0.1,
-    },
-  },
-};
-
-
-const navItems = [
-  {
-    title: "Blogs",
-    path: "/blogs",
-  },
-  {
-    title: "Quizzes",
-    path: "/quiz",
-  },
-  {
-    title: "Projects",
-    path: "/projects",
-  },
-  {
-    title: "Community",
-    path: "/community",
-  },
-];
-
-
-
-export default function BlogListingGridPage() {
-  const [searchTerm, setSearchTerm] = useState("");
-  const [isMenuOpen, setIsMenuOpen] = useState(false);
-
-  const router = useRouter();
-
-  const filteredPosts = blogPosts.filter(
-    (post) =>
-      post.title.toLowerCase().includes(searchTerm.toLowerCase()) ||
-      post.description.toLowerCase().includes(searchTerm.toLowerCase()) ||
-      post.category.toLowerCase().includes(searchTerm.toLowerCase())
-  );
-
-
+const BlogsPage = () => {
   return (
-    <div className="min-h-screen bg-blue-950 text-white">
-      <header className="container mx-auto py-6 px-4 flex justify-between items-center">
-        <Link href="/" className="flex items-center space-x-2">
-          <Code2 className="h-8 w-8 text-blue-400" />
-          <span className="text-2xl font-bold bg-clip-text text-transparent bg-gradient-to-r from-blue-400 to-purple-400">
-            GoBoolean
-          </span>
-        </Link>
-        <nav className="hidden md:block">
-          <ul className="flex space-x-6">
-            {navItems.map((item) => (
-              <li key={item.title}>
-                <Link
-                  href={`${item.path}`}
-                  className="hover:text-blue-300 transition-colors"
-                >
-                  {item.title}
-                </Link>
-              </li>
-            ))}
-          </ul>
-        </nav>
-
-        <Sheet open={isMenuOpen} onOpenChange={setIsMenuOpen}>
-          <SheetTrigger asChild>
-            <Button
-              variant="ghost"
-              size="icon"
-              className="md:hidden"
-              aria-label="Open menu"
-            >
-              <Menu className="h-6 w-6" />
-            </Button>
-          </SheetTrigger>
-          <SheetContent
-            side="right"
-            className="w-[300px] sm:w-[400px] bg-blue-900"
-          >
-            <SheetHeader>
-              <SheetTitle className="text-2xl font-bold text-blue-300">
-                Menu
-              </SheetTitle>
-            </SheetHeader>
-            <nav className="mt-6">
-              <ul className="space-y-4">
-                {navItems.map((item) => (
-                  <li key={item.title}>
-                    <Link
-                      href={`${item.path}`}
-                      className="block py-2 px-4 text-lg text-blue-100 hover:bg-blue-800 rounded transition-colors"
-                      onClick={() => setIsMenuOpen(false)}
-                    >
-                      {item.title}
-                    </Link>
-                  </li>
-                ))}
-              </ul>
-            </nav>
-          </SheetContent>
-        </Sheet>
-      </header>
-
-      <main className="container mx-auto py-12 px-4">
-        <motion.div
-          initial="initial"
-          animate="animate"
-          variants={stagger}
-          className="space-y-8"
-        >
-          <motion.h1
-            className="text-4xl font-bold text-center bg-clip-text text-transparent bg-gradient-to-r from-blue-400 to-purple-400"
-            variants={fadeIn}
-          >
-            GoBoolean Blog
-          </motion.h1>
-          <motion.div variants={fadeIn} className="max-w-md mx-auto">
-            <div className="relative">
-              <Search className="absolute left-3 top-1/2 transform -translate-y-1/2 text-blue-300" />
-              <Input
-                type="text"
-                placeholder="Search blog posts..."
-                value={searchTerm}
-                onChange={(e) => setSearchTerm(e.target.value)}
-                className="pl-10 bg-white/10 border-white/20 text-white placeholder:text-white/50"
-              />
-            </div>
-          </motion.div>
-          <motion.div
-            variants={fadeIn}
-            className="grid gap-6 md:grid-cols-2 lg:grid-cols-3"
-          >
-            {filteredPosts.map((post) => (
-              <Card
-                key={post.id}
-                // onClick={() => window.open(`/blogs/${post.slug}`, "_blank")}
-                // onClick={() => window.open(`/blogs/blog-one`, "_blank")}
-                // onClick={() => router.push(`/blogs/blog-one`)}
-                onClick={() => router.push(`/blogs/${post.slug}`)}
-                className="bg-white/10 border-white/20 hover:bg-white/20 transition-all duration-300"
-              >
-                <CardHeader>
-                  <CardTitle className="text-blue-300">{post.title}</CardTitle>
-                  <CardDescription className="text-blue-200">
-                    {post.category} • {post.readTime}
-                  </CardDescription>
-                </CardHeader>
-                <CardContent>
-                  <p className="text-blue-100">{post.description}</p>
-                </CardContent>
-                <CardFooter className="flex justify-between items-center">
-                  <div className="text-sm text-blue-200">
-                    By {post.author} on{" "}
-                    {new Date(post.date).toLocaleDateString()}
-                  </div>
-                  <Button
-                    variant="ghost"
-                    className="text-blue-300 hover:text-blue-100 hover:bg-blue-500/20"
-                  >
-                    Read More <ChevronRight className="ml-2 h-4 w-4" />
-                  </Button>
-                </CardFooter>
-              </Card>
-            ))}
-          </motion.div>
-        </motion.div>
-      </main>
-
-      <footer className="bg-blue-950/50 py-10 mt-12">
-        <div className="container mx-auto text-center text-sm text-blue-300">
-          © {new Date().getFullYear()} goboolean.in All rights reserved.
-        </div>
-      </footer>
+    <div>
+        <BlogsComingSoon />
     </div>
-  );
+  )
 }
+
+export default BlogsPage
