+++ components/Footer.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Footer = () => {
+  return (
+    <div>
+        Footer content goes here
+    </div>
+  )
+}
+
+export default Footer
